app/requirements.txt
django>=2.2.4
mysqlclient
django-debug-toolbar
faker
gunicorn
pytest
pytest-django
pytest-cov
setuptools
ipdb
djangorestframework
django-rest-auth
Pillow
djangorestframework_simplejwt
celery==4.3.0
redis==3.2.1
django-celery-beat
sentry-sdk==0.11.2
pyspy